+++ examples/01-choco/aws_ppr.tf
@@ -1,23 +1,23 @@
-module "context_choco_aws_ppr" {
-  source = "../.."
+# module "context_choco_aws_ppr" {
+#   source = "../.."
 
-  #Common
-  cloud       = "aws"
-  environment = "ppr"
-  region      = "northeurope"
+#   #Common
+#   cloud       = "aws"
+#   environment = "ppr"
+#   region      = "northeurope"
 
-  #General
-  department = "legal"
-  project    = "cocoa"
-  tags = {
-    OwnerContact = "wonka@choco.me"
-    OwnerOU      = "choco"
-    Sweetness    = "very"
-  }
-  custom_tags = {
-    HQ = "Brussels"
-  }
-}
-output "context_choco_aws_ppr" {
-  value = module.context_choco_aws_ppr.context
-}
+#   #General
+#   department = "legal"
+#   project    = "cocoa"
+#   tags = {
+#     OwnerContact = "wonka@choco.me"
+#     OwnerOU      = "choco"
+#     Sweetness    = "very"
+#   }
+#   custom_tags = {
+#     HQ = "Brussels"
+#   }
+# }
+# output "context_choco_aws_ppr" {
+#   value = module.context_choco_aws_ppr.context
+# }