+++ modules/choco/outputs.tf
@@ -1,4 +1,4 @@
-#context
+#Context
 output "context" {
   value = local.context
 }