+++ modules/choco/variables.tf
@@ -21,80 +21,31 @@ variable "cloud" {
 variable "environment" {
   type        = string
   description = "Environment - place in the SDLC lifecycle. Valid values for environment: dev, tst, ppr, prd."
-  default = null
   validation {
-    condition     = var.environment == null || anytrue([  
-                      for val in ["dev", "tst", "ppr", "prd"]: val == var.environment 
-                    ])  
+    condition     = contains(["dev", "tst", "ppr", "prd"], var.environment)
     error_message = "Valid values for environment: dev, tst, ppr, prd"
   }
 }
 
-variable "environment_azure" {
-  type        = string
-  description = "Azure Environment - place in the SDLC lifecycle. Valid values for environment: dev, tst, ppr, prd."
-  default = null
-  validation {
-    condition     = var.environment_azure == null || anytrue([  
-                      for val in ["dev", "tst", "ppr", "prd"]: val == var.environment_azure  
-                    ])
-    error_message = "Valid values for environment: dev, tst, ppr, prd"
-  }
-}
-
-variable "environment_aws" {
-  type        = string
-  description = "AWS Environment - place in the SDLC lifecycle. Valid values for environment: dev, tst, ppr, prd."
-  default = null
-  validation {
-    condition     = var.environment_aws == null || anytrue([  
-                      for val in ["dev", "tst", "ppr", "prd"]: val == var.environment_aws
-                    ])
-    error_message = "Valid values for environment: dev, tst, ppr, prd"
-  }
-}
-
-variable "environment_gcp" {
-  type        = string
-  description = "GCP Environment - place in the SDLC lifecycle. Valid values for environment: dev, tst, ppr, prd."
-  default = null
-  validation {
-    condition     = var.environment_gcp == null || anytrue([  
-                      for val in ["dev", "tst", "ppr", "prd"]: val == var.environment_gcp
-                    ])
-    error_message = "Valid values for environment: dev, tst, ppr, prd"
-  }
-}
-
-variable "environment_oci" {
-  type        = string
-  description = "OCI Environment - place in the SDLC lifecycle. Valid values for environment: dev, tst, ppr, prd."
-  default = null
-  validation {
-    condition     = var.environment_oci == null || anytrue([  
-                      for val in ["dev", "tst", "ppr", "prd"]: val == var.environment_oci
-                    ])
-    error_message = "Valid values for environment: dev, tst, ppr, prd"
-  }
-}
-
-variable "environment_onprem" {
-  type        = string
-  description = "OnPrem Environment - place in the SDLC lifecycle. Valid values for environment: frankfurt."
-  default = null
-  validation {
-    condition     = var.environment_onprem == null || anytrue([
-                      for val in ["frankfurt"]: val == var.environment_onprem
-                    ])
-    error_message = "Valid values for environment: frankfurt"
-  }
-}
 #Region
 variable "region" {
   type        = string
   description = "Cloud Region. Valid values for the region: frankfurt, westeurope, northeurope and germanywestcentral."
   validation {
-    condition     = contains(["frankfurt", "westeurope", "northeurope", "germanywestcentral"], var.region)
-    error_message = "Valid values for the region: frankfurt, westeurope, northeurope and germanywestcentral"
+    condition = contains([
+      "eastus2", "centralus", "westus2", "northeurope", "westeurope", "germanywestcentral",                #Azure
+      "us-east-1", "us-west-1", "us-west-2", "eu-west-1", "eu-central-1", "eu-north-1",                    #AWS
+      "us-central1", "us-east1", "us-west1", "europe-west1", "europe-west2", "europe-west3",               #GCP
+      "us-ashburn-1", "us-phoenix-1 ", "us-sanjose-1 ", "eu-frankfurt-1", "eu-amsterdam-1", "eu-london-1", #OCI
+      "us-central", "us-east", "us-west", "eu-central", "eu-west", "eu-north"                              #OnPrem
+    ], var.region)
+    error_message = <<EOT
+    Valid values for the regions are:
+    azure  = ["eastus2", "centralus", "westus2", "northeurope", "westeurope", "germanywestcentral"]
+    aws    = ["us-east-1", "us-west-1", "us-west-2", "eu-west-1", "eu-central-1", "eu-north-1"]
+    gcp    = ["us-central1", "us-east1", "us-west1", "europe-west1", "europe-west2", "europe-west3"]
+    oci    = ["us-ashburn-1", "us-phoenix-1 ", "us-sanjose-1 ", "eu-frankfurt-1", "eu-amsterdam-1", "eu-london-1"]
+    onprem = ["us-central", "us-east", "us-west", "eu-central", "eu-west", "eu-north"].
+    EOT
   }
 }