+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # Change Log
 
-## v0.3.0 [current]
+## v0.5.0 [current]
+FEATURES:
+- Added GCP data
+
+ENHANCEMENTS:
+- Updated logic for handling different regions for each cloud
+
+BUG FIXES:
+- Fixed bug, when you needed to set cloud foreign regions to data sets
+
+## v0.4.0 [deprecated]
+ENHANCEMENTS:
+- Added cloud specific environment variables
+
+## v0.3.0
 FEATURES:
 - Added the Candy Shop